Add threaded conversation viewing
This allows a user to view a conversation thread on twitter. It is initiated
with "threads <ID>" (i.e. "threads 3f") and by default a max of 5 tweets is
shown. This can be configured via the "conversation_max_tweets" setting. The
resulting conversation is fetched asynchronously and delivered as PRIVMSGs.